==> 1.0 Introduction - Data.ipynb <==
# Predicting the impact of Markdowns on Weekly Sales in a Worldclass Retailer

## Introduction

In this project, we will be using data sets that reports the weekly sales and other vital information for a retail giant in the United States. The dataset include data on holidays and select major events that come up once a year for each department in the reatil store. 

It is the objective of the project to anaylise the data set so as to enable store executives make strategic decisions which would ultimately affect the bottom line. It is often suggested that markdowns do affect sales, so this project also sets out to test it as a hypothesis and subsequently predict which departments will be affected and to what extent.

## Data
The data sets contain historical sales data for 45 stores located in different regions in the United States. Each store contains a number of departments. The company also runs several promotional markdown events throughout the year. These markdowns precede prominent holidays, the four largest of which are the Super Bowl, Labor Day, Thanksgiving, and Christmas. The weeks including these holidays are weighted five times higher in the evaluation than non-holiday weeks. 

One of the major challenges of modeling retail data is the need to make decisions based on limited history.  

The data sets are in three tranches. The first is a 8191 x 12 features data set. The 12 columns are as follows:

Contains additional data related to the store, department, and regional activity for the given dates.

1. Store - The store number
2. Date - The date for the day of the week
3. Temperature - The average temperature in the region
4. Fuel_Price - The cost of fuel in the region
5. MarkDown1-5 - These are anonymized data related to promotional markdowns. Take note that MarkDown data is only available after Nov 2011, and is not available for all stores all the time. Any missing value is marked with an NA.
6. CPI - The consumer price index
7. Unemployment - The unemployment rate
8. IsHoliday - This is a boolean variable that determines whether the week is a special holiday week

The next data set tranche is a 46 x 3 Stores data set. This data set contain anonymized information about the 45 stores, indicating the type and size of store.

The last data set is the 421571 x 5 Sales data set. This data set contains historical sales data, which covers weekly sales from 2010-02-05 to 2012-11-01. Within this data set we will find the following columns:

1. Store - The store number
2. Dept - The department number
3. Date - The date for the day of the week
4. Weekly_Sales - The weekly sales for the given department in the given store
5. IsHoliday - This is a boolean variable that determines whether the week is a special holiday week

### Loading the data sets

#### Feature data set
We will start by loading the retail feature data set. The data set is called Features data set.csv. 

# import the python required libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# load the Features data set.csv
df_features = pd.read_csv('Features data set.csv')
df_features.head()

We observe that the data set contains missing values which correspond primarily to unavailable MarkDown data. These missing values will be handled in the section on DATA WRANGLING.

#### Stores Data Set
The next data set that would be loaded for use in this study is the Stores data set. As highlighted above, the data set consists of 46 rows representing the 45 stores + the index and 3 columns comprising the store number, type and size. This data set will be merged with the Features data set as part of the master data set for the project. The procedure will be discussed in the DATA WRANGLING section.

# load the stores data-set.csv
df_stores = pd.read_csv('stores data-set.csv')
df_stores.head()

#### Sales Data Set

The last data set to be loaded as part of the project is the sales data "set data-set.csv". The sales data set is a 421571 x 5 data tab that contains historical sales data for 98 departments of the retail store. It also has a column for denoting whether the weekly sale was taken in a holiday or not.   

# load the sales data-set.csv
df_sales = pd.read_csv('sales data-set.csv')
df_sales.head()

### Data Wrangling and Conversion

In this section, we will clean up the data and perform data conversion so that we will have master data set that is suitable for the analysis.

#### Merging the features data set and the stores data set

The first step in creating a master data set for analysis is to merge the features data set and the stores data set. This will create a data set that will contain the store type and size as features.  We will call this data set df_master1. 

# Create a master data set comprising a combination of the feature data set and the stores data set
df_master1 = df_features.merge(df_stores,how='left', left_on='Store', right_on='Store')
df_master1.head()

#### Preparing sales data set for incorporation to the master data set

The sales data set contains 421571 rows which comprises weekly sales data for 98 departments per store number and for each week from January 2010. In order to make the data set consistent with the other data sets so that they could easily be merged to create a master data set, the departments need to be moved to the columns. The first step to achieve this is by ensuring that the Date column is converted to datetime. 

# Convert 'Date' in df_sales to datetime
df_sales['Date'] =pd.to_datetime(df_sales.Date)

The next step is to use the unstack method to create a sales table with the departments as columns rather than rows. This creates a 6435 rows × 81 columns table.  

df_sales1 = df_sales.set_index(['Store','Date','Dept'])['Weekly_Sales'].unstack()
df_sales1.head()

In order to export the data as excel spreadsheet to view the data set for consistency, we will need to import the necessary libraries

from pandas import ExcelWriter
from pandas import ExcelFile

writer = pd.ExcelWriter('sales_output.xlsx')

df_sales1.to_excel(writer,'Sheet1')

writer.save()

We will now load the sales_output.xlsx as dataframe. We will call it df_sales2

# load the output.xls
df_sales2 = pd.read_excel('sales_output.xlsx', sheetname='Sheet1')
df_sales2.head()

#### Creating the Master Data Set for Analysis

We will now create a master data set for analysis by merging df_sales2 and df_master1. This master data set shall be called df_master.

df_master = pd.concat([df_master1, df_sales2], axis=1)
df_master.head()

#### Handling Missing Values

The missing values in the first few rows will be filled using the backfill method while the remainder will be filled using the interpolate method. 

Let us fill the first few columns:

df_master = df_master.fillna(method='bfill')
df_master.head() 

And then we will use the interpolate method to fill the remaining missing values

df_master = df_master.interpolate()
df_master.tail()

#### Delete Duplicate Columns

We now drop the duplicate store and date columns. 

# Drop duplicate columns
df_master.drop(df_master.columns[[14,15]], axis=1, inplace=True)

#### Check to see if there are any missing values in our data set

We will now check to see if there are still missing values in our data set.

# Check to see if there are any missing values in our data set
df_master.isnull().any()

#### Renaming certain columns for better readability

We will rename some of the columns for better readability

df_master.columns = ['Temperature','Fuel_Price', 'MarkDown1', 'MarkDown2', 'MarkDown3', 'MarkDown4', 'MarkDown5', 'CPI',
                      'Unemployment', 'IsHoliday', 'Type', 'Size', 'Jewelry', 'Pets', 'TV_Video', 'Cell_Phones', 
                      'Pharmaceutical ', 'Health_beauty', 'Toy ', 'Home_others', 'Kitchen', 'Bedding', 'Bathroom', 
                      'Office_supplies ', 'School_Supplies', 'Home_Office', 'Craft_general', 'Floral', 'Beading', 
                      'Paint', 'Framing', 'outdoor', 'Auto', 'School_Uniforms', 'Baby_Toddlers_Clothing', 
                      'Baby_Kids_Shoes', 'Clearance_Clothings', 'Boys_Clothing', 'Girls_Clothing', 'Women_Clothing', 
                      'Intimates_Sleepwears', 'Men_Clothings', 'Precious_Metals', 'Active_Wear', 'Adult_Shoes', 
                      'Bags_Accessories', 'Sportswear', 'Computer', 'Music', 'Luggage', 'Food', 'Fruit', 'Grocery', 
                      'Laundry', 'IPad_Tablets', 'Heating_Cooling', 'Swim_Shop', 'Gift_cards', 'Baby_Essentials', 
                      'Cribs', 'Car_Seats', 'Strollers', 'Bikes', 'Photo', 'Household_Essentials', 'Air_Quality', 
                      'Light_bulbs', 'Gardening', 'Building_Materials', 'Hardware', 'Electrical', 'Home_Safety', 
                      'Tools', 'Teen_Room', 'Kids_Room', 'Lighting', 'Home_Decor', 'Mattresses', 'Furniture', 'Storage', 'Appliances', 
                      'Pioneer_Woman', 'Computer_Software', 'Books', 'Musical_Instruments', 'Star_Wars', 'Movies_TV', 'Video_Games', 
                      'Portable_Audios', 'Cameras_Camcoders', 'Auto_Electronics', 'Wearable_Tech', 'Smart_homes']
                
                                    

df_master.head()

#### Saving the master data set as excel spreadsheet

Finally, we save master data set as excel spreadsheet so that it could be easily imported in other notebooks for further analysis. 

writer = pd.ExcelWriter('master_dataset.xlsx')
df_master.to_excel(writer,'Sheet1')
writer.save()
