==> 8.0 Learning and Predicting.ipynb <==
### 8.0.1 Support Vector Classification
In the case of our retail dataset, the task is to predict, Weekly Sales, which the MarkDowns represent. We have samples of each of the 5 markdowns on which we fit an estimator to be able to predict the classes to which unseen samples belong. 

We are going to derive an estimator for classification which is a Python object that implements the methods fit(X, y) and predict(T). The estimator we are using is the class sklearn.svm.SVC that implements support vector classification.

from sklearn import svm

clf = svm.SVC(gamma=0.001, C=100.)

#### 8.0.1.1 Choosing the parameters of the model
The constructor of the estimator in 8.0.1 takes as arguments the parameters of the model.

##### 8.0.1.1.1 Manual Setting of Gamma
First we set the value of gamma manually. The estimator instance clf (which is a classifier) is fitted to the model, that is, we let it learn from the model.

# libraries
#%matplotlib notebook

import pandas as pd
import numpy as np

import matplotlib
import seaborn
import matplotlib.dates as md
from matplotlib import pyplot as plt

from sklearn import preprocessing
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from sklearn.covariance import EllipticEnvelope
from pyemma import msm
from sklearn.ensemble import IsolationForest
from sklearn.svm import OneClassSVM

# load the master_dataset.xls
df = pd.read_excel('master_dataset.xlsx', sheetname='Sheet1')

print(df.info())

#### MarkDown 1 Prediction
Here we predict the behavior of MarkDown 1 using the sales of other materials and other parameters.

# Use MarkDown1 sales as a target for prediction. 
df_MarkDown1 = df['MarkDown1']

Split Markdown1 into Low-Low Sales represented by the number 0, Low Sales represented by the number 1,  Average Sales represented by the number 2, Above Average Sales represented by the number 3 and High Sales represented by the number 4 / LLS, LS, AS, AAS, HS

def MarkDown1_Split(x):
    if x < 4444:
        return 0
    elif 4444< x < 8888:
        return 1
    elif 8888< x < 13332:
        return 2
    elif 13332 < x < 17776:
        return 3
    else: return 4

# Descritize MarkDown_Score
df_MarkDown1["MarkDown_Score"] = df_MarkDown1.apply(MarkDown1_Split)

print (df_MarkDown1["MarkDown_Score"].head())

#Drop none-float columns and the MarkDowns that would be used for prediction
df.drop(['Store', 'Date', 'IsHoliday', 'Type', 'MarkDown1', 'MarkDown2', 'MarkDown3', 'MarkDown4', 'MarkDown5'], axis=1, inplace=True)

As a training set, we would use all the values of our dataset apart from the last one. We select this training set with the [:-1] Python syntax, which produces a new array that contains all but the last entry of digits.data

clf.fit(df[:-1], df_MarkDown1["MarkDown_Score"][:-1])

Now we can predict new values, in particular, we will ask to the classifier what is the value of our last data point in the retail dataset, which we have not used to train the classifier:

clf.predict(df[-1:])

This has equivalent of Low MarkDown1 Sale.

##### 8.0.1.1.2 Automatic Setting of Gamma¶
It is possible to automatically find good values for the parameters. In order to do this, we will be using tools such as grid search and cross validation.
