==> 9.0 Variance Inflation Factor.ipynb <==
### 9.0.1 Collinearity
In this section, we will be exploring the degree of correlation between the predictor variables in our retail sales dataset. That is, we will be testing the models collinearity. For two variables that are collinear, we expect that they should contain similar information about the variance within the given dataset. 

To detect the collinearity, we will create a correlation matrix and find variables with large absolute values.

### 9.0.2 Multicollinearity
Detection of multicollinearity is follows a more complicated procedure. This is because it emerges when three or more variables which are highly correlated are included in the model. It can also emerge when isolated pairs of variables are not collinear.

In order to test multicollinearity we will use Variance Infaltion Factor (VIF). 

### 9.0.3 Variance Inflation Factor
The variance inflation factor (VIF) is a measure of colinearity among predictor variables within a multiple regression. It is calculated by taking the ration of the variance of all a given model's beta if it were fit alone. 

### 9.0.4 Steps for Implementing VIF

1. Run a multiple regression
2. Calculate the VIF Factors
3. Inspect the factors for each predictor variable: If the VIF is between 5-10, mul;ticollinearity is likely present and we would consider dropping the variable.

Let us import the libraries required for analysis.

#imports
import pandas as pd
import numpy as np
from patsy import dmatrices
import statsmodels.api as sm
from statsmodels.stats.outliers_influence import variance_inflation_factor

### 9.0.5 Data Import
We will now import the dataset.

# load the master_dataset.xls
df = pd.read_excel('master_dataset.xlsx', sheetname='Sheet1')

### 9.0.6 Data Manipulation
We will drop all the non-numeric columns

# Drop Non-numeric cells
df.dropna()

# Drop Non-Numeric Columns
df = df._get_numeric_data()

df.head()

Let us use MarkDown1 as the target variable and the others as predictor variable and then run multiple regression. Firstly, let us create a duplicate dataframe and then we will drop the column "MarkDown1" which we will use as our target varibale from the duplicate dataframe.

target = df['MarkDown1']

# Duplicate dataframe and assign the name df_dup to it
df_dup = df

# Drop "Markdown1"
df_dup.drop(['MarkDown1'],  axis = 1, inplace=True)

### 9.0.7 Numpy Array
We will create two numpy arrays X, representing the predictor variables and y representing the target variables:

#Numpy array
X = df_dup.values
y = target.values

### 9.0.8 Reshape Array
We will now reshape the arrays in preparation for regression analysis

# Print the dimensions of X and y before reshaping
print("Dimensions of y before reshaping: {}".format(y.shape))
print("Dimensions of X before reshaping: {}".format(X.shape))

# Reshape  the dimension of y
y = y.reshape(-1, 1)

### 9.0.9 Multiple Regression

## fit a OLS model with intercept on the data set
X = sm.add_constant(X)


We will run linear regression models in SKLearn. SKLearn is pretty much the golden standard when it comes to machine learning in Python. In order to use linear regression, we need to import it:

from sklearn import linear_model

Then we’ll fit a model:

lm = linear_model.LinearRegression()

model = lm.fit(X,y)

The lm.fit() function fits a linear model. We want to use the model to make predictions, so we’ll use lm.predict():

predictions = lm.predict(X)

The print function would print the first 5 predictions for y (I didn’t print the entire list to “save room”. Removing [0:5] would print the entire list):

print((predictions)[0:5])

Remember, lm.predict() predicts the y (dependent variable) using the linear model we fitted. We use built-in functions to return the score, the coefficients and the estimated intercepts

lm.score(X,y)

lm.coef_

lm.intercept_

These are all (estimated/predicted) parts of the multiple regression equation 

### 9.0.10 Generating VIF

# load the master_dataset.xls
df = pd.read_excel('master_dataset.xlsx', sheetname='Sheet1')

#subset the dataframe
df = df[['MarkDown1','MarkDown2','MarkDown3', 'MarkDown4', 'MarkDown5', 'CPI', 'Unemployment']].dropna() 

#### 9.0.10.1 Run Multiple Regression

# Run Multiple Regression
#%%capture
#gather features
features = "+".join(df.columns)


# get y and X dataframes based on this regression:
y, X = dmatrices('MarkDown1 ~' + features, df, return_type='dataframe')

#### 9.0.10.2 Calculate VIF Factors

# For each X, calculate VIF and save in dataframe
vif = pd.DataFrame()
vif["VIF Factor"] = [variance_inflation_factor(X.values, i) for i in range(X.shape[1])]
vif["features"] = X.columns

vif.round(1)

# load the master_dataset.xls
df = pd.read_excel('master_dataset.xlsx', sheetname='Sheet1')

# Drop Non-Numeric Columns
df = df._get_numeric_data()

df.keys()

df[['Toy']] = df[['Toy ']]

df[['Pharmaceutical']] = df[['Pharmaceutical ']]

df[['Office_supplies']] = df[['Office_supplies ']]

df = df[['MarkDown1', 'Store', 'Temperature', 'Fuel_Price', 'MarkDown1', 'MarkDown2', 'MarkDown3', 'MarkDown4', 'MarkDown5', 
         'CPI', 'Unemployment', 'Size', 'Jewelry', 'Pets', 'TV_Video', 'Cell_Phones', 'Pharmaceutical', 'Health_beauty', 'Toy',
         'Home_others', 'Kitchen', 'Bedding', 'Bathroom', 'Office_supplies', 'School_Supplies', 'Home_Office', 'Craft_general',
         'Floral','Beading', 'Paint', 'Framing', 'outdoor', 'Auto', 'School_Uniforms', 'Baby_Toddlers_Clothing', 'Baby_Kids_Shoes',
         'Clearance_Clothings', 'Boys_Clothing', 'Girls_Clothing', 'Women_Clothing', 'Intimates_Sleepwears', 'Men_Clothings',
         'Precious_Metals', 'Active_Wear', 'Adult_Shoes', 'Bags_Accessories', 'Sportswear', 'Computer', 'Music', 'Luggage',
         'Food', 'Fruit', 'Grocery', 'Laundry', 'IPad_Tablets', 'Heating_Cooling', 'Swim_Shop', 'Gift_cards', 'Baby_Essentials',
         'Cribs', 'Car_Seats', 'Strollers', 'Bikes', 'Photo', 'Household_Essentials', 'Air_Quality', 'Light_bulbs', 'Gardening',
         'Building_Materials', 'Hardware', 'Electrical', 'Home_Safety', 'Tools', 'Teen_Room', 'Kids_Room', 'Lighting', 
         'Home_Decor', 'Mattresses', 'Furniture', 'Storage', 'Appliances', 'Pioneer_Woman', 'Computer_Software', 'Books', 
         'Musical_Instruments', 'Star_Wars', 'Movies_TV', 'Video_Games', 'Portable_Audios', 'Cameras_Camcoders', 
         'Auto_Electronics', 'Wearable_Tech', 'Smart_homes']].dropna()

# Run Multiple Regression
#%%capture
#gather features
features = "+".join(df.columns)

# get y and X dataframes based on this regression:
y, X = dmatrices('MarkDown1 ~' + features, df, return_type='dataframe')

# For each X, calculate VIF and save in dataframe
vif = pd.DataFrame()
vif["VIF Factor"] = [variance_inflation_factor(X.values, i) for i in range(X.shape[1])]
vif["features"] = X.columns

vif.round(1)

vif.round(1).iloc[1:29]

vif.round(1).iloc[29:59]

vif.round(1).iloc[59:79]

vif.round(1).iloc[79:92]

The features that have VIF less than 5 are: 
1. Temperature
2. Fuel_Price
3. MarkDown2
4. MarkDown3
5. MarkDown4
6. MarkDown5
7. CPI
8. Unemployment
9. Size
10. Jewelry
11.	Bathroom
12. Beading
13. Paint
14. Sportswear
15.	Music
16.	Fruit
17.	Laundry
18. Heating_Cooling
19.	Gift_cards
20.	Baby_Essentials
21. Car_Seats
22.	Strollers
23. Photo
24. Air_Quality
25. Light_bulbs
26. Gardening
27. Building_Materials
28.	Kids_Room
29. Lighting

We will combine the features that are related with high VIF (>5.0) so as to create new features. These new features are:

* Home-items
* Electronic-items
* Health-items
* Kids-items
* Office-items
* Auto-items
* Wears-items
* Food-items

These newly created features will be used as target variables. They will be regressed against the 29 features that have less than VIF of 5. 

df['Home-items'] = (120.100000*df['Pets'] + 6.100000*df['Home_others'] + 8.400000*df['Kitchen'] + 23.700000*df['Bedding'] + 
                    4.900000*df['Bathroom'] + 32.900000*df['Floral'] + 30.2*df['Framing'] + 20.6*df['outdoor'] + 
                    15.7*df['Household_Essentials'] + 11.4*df['Home_Safety'] + 17.6*df['Home_Decor'] + 25.1*df['Mattresses'] + 
                    45.5000*df['Furniture'] + 17.3000*df['Storage'])/(120.100000 + 6.10000 + 8.40000 + 23.700000 + 4.900000 + 
                                                                      32.900000 + 30.2 + 20.6 + 15.7 + 11.4 + 17.6 + 25.1 + 
                                                                      45.5000 + 17.3000)

df['Electronic-items'] = (8.200000*df['TV_Video'] + 76.300000*df['Cell_Phones'] + 11.7*df['Computer'] + 17.3*df['IPad_Tablets']
                         + 6.7*df['Electrical'] + 11.3*df['Appliances'] + 7.7*df['Computer_Software'] + 
                          60.7*df['Musical_Instruments'] + 116.6* df['Star_Wars'] + 21.2*df['Movies_TV'] + 32.8*df['Video_Games']
                         + 89.0*df['Portable_Audios'] + 6.8*df['Cameras_Camcoders'] + 46.2*df['Auto_Electronics']
                         )/(8.200000 + 76.300000 + 11.7 + 17.3 + 6.7 + 11.3 + 7.7 + 21.2 + 32.8 + 89.0)

df['Health-items'] = (18.900000*df['Pharmaceutical'] + 9.700000*df['Health_beauty'])/(18.900000 + 9.700000)

df['Kids-items'] = (19.600000*df['Toy'] + 48.6*df['School_Uniforms'] + 26.9*df['Baby_Toddlers_Clothing'] + 
                    16.6*df['Baby_Kids_Shoes'] + 6.7*df['Cribs'] + 11.4*df['Bikes'] + 31.9*df['Teen_Room']
                   )/(19.600000 + 48.6 + 26.9 + 11.4 + 31.9)

df['Office-items'] = (10.800000*df['Office_supplies'] + 88.600000*df['School_Supplies'] + 44.600000*df['Home_Office'] +
                     11.000000*df['Craft_general'] + 134.3*df['Books'])/(10.800000 + 88.600000 + 44.600000 + 11.000000 + 134.3)

df['Auto-items'] = (29.4*df['Auto'] + 14.0*df['Precious_Metals'] + 7.3*df['Hardware'] + 73 + 
                    11.4*df['Tools'])/(29.4 + 14.0 + 7.3 + 11.4)

df['Wears-items'] = (23.0*df['Clearance_Clothings'] + 18.7*df['Boys_Clothing'] + 9.1*df['Girls_Clothing'] + 
                        21.4*df['Women_Clothing'] + 13.9* df['Intimates_Sleepwears'] + 11.3*df['Men_Clothings'] + 
                     34.5*df['Active_Wear'] + 18.7*df['Adult_Shoes'] + 8.4*df['Bags_Accessories'] + 6.7*df['Luggage'] + 
                     44.2*df['Swim_Shop'] + 5.7*df['Pioneer_Woman']
                       )/(23.0 + 18.7 + 9.1 + 13.9 + 18.7 + 8.4 + 44.2 + 5.7)

df['Food-items'] = (56.3*df['Food'] + 17.3*df['Grocery'])/(56.3 + 17.3)
