==> 18.0 Regression-3.ipynb <==
## 18.0 Introduction
We will explore further regression of the predictors selected after calculation of the VIFs and the target variables: 
* Transport (Auto)-Items
* Fashion (Wears)-Items and
* Food-Items

We will use statsmodel and sklearn for multiple regression. 

### 18.0.1 Python Libraries
In order to start, we will import necessary libraries. 

#import Libraries
import pandas as pd
import numpy as np
import seaborn as sns; sns.set()
import matplotlib
import matplotlib.pyplot as plt
import statsmodels.api as sm
from sklearn import linear_model
%matplotlib inline  

## 18.1 LINEAR REGRESSION IMPLEMENTING VIF RESULT

### 18.1.0 TARGET - Health-items 
Objective: Predict Health Items sales performance 

# Read the file 'master_dataset.xlsx' into a DataFrame df using the read_xls() function.
df = pd.read_excel('master_dataset.xlsx', sheetname='Sheet1')

After importing the master dataset 'master_dataset.xlsx', we create features dataframe using the columns with VIF<5

# generating the 29 predictor features
features = df[['Temperature', 'Fuel_Price', 'MarkDown1', 'MarkDown2', 'MarkDown3', 'MarkDown4', 'MarkDown5', 'CPI', 'Unemployment',
               'Size','Jewelry', 'Bathroom', 'Beading', 'Paint', 'Sportswear', 'Music', 'Fruit', 'Laundry', 'Heating_Cooling',
               'Gift_cards', 'Baby_Essentials', 'Car_Seats', 'Strollers', 'Photo', 'Air_Quality', 'Light_bulbs', 'Gardening',
               'Building_Materials', 'Kids_Room', 'Lighting']]

We now create a column for health items using the VIFs as weighting factors:  

# Target 
df['Auto-items'] = (29.4*df['Auto'] + 14.0*df['Precious_Metals'] + 7.3*df['Hardware'] + 73 + 
                    11.4*df['Tools'])/(29.4 + 14.0 + 7.3 + 11.4)

#### 18.1.1.1 Linear Regression Using Statsmodel
We will proceed to run multiple regression using Statsmodel. First, we will fit and summarize the OLS model with zero intercept.

##### 18.1.1.1.1 Linear Regression Using Statsmodel; without intercept

# Fit and summarize OLS model
mod = sm.OLS(df['Auto-items'], features)

res = mod.fit()

print(res.summary())

#### 18.1.1.1.2 Linear Regression Using Statsmodel¶ - with intercept

# fit a OLS model with intercept on the data set
X = sm.add_constant(features)

# Fit and summarize OLS model
mod = sm.OLS(df['Auto-items'], X)

res = mod.fit()

print(res.summary())

#### 18.1.1.2 Linear Regression Using SKLearn

We will run linear regression models in SKLearn.

Then we’ll create numpy arrays of the features and target, reshape the target and fit a model:

#create numpy array
X = features.values
y = df['Auto-items'].values

#Reshape y
y = y.reshape(-1,1)

#fit regression model
lm = linear_model.LinearRegression()
model = lm.fit(X,y)

The lm.fit() function fits a linear model. We want to use the model to make predictions, so we’ll use lm.predict():

#run prediction 
predictions = lm.predict(X)

The print function would print the first 5 predictions for y (I didn’t print the entire list to “save room”. Removing [0:5] would print the entire list):

print((predictions)[0:5])

# generate the score
lm.score(X,y)

lm.coef_

lm.intercept_

### 18.2.1 TARGET - Fashion-items
Objective: Predict Fashion Items sales performance

df['Wears-items'] = (23.0*df['Clearance_Clothings'] + 18.7*df['Boys_Clothing'] + 9.1*df['Girls_Clothing'] + 
                        21.4*df['Women_Clothing'] + 13.9* df['Intimates_Sleepwears'] + 11.3*df['Men_Clothings'] + 
                     34.5*df['Active_Wear'] + 18.7*df['Adult_Shoes'] + 8.4*df['Bags_Accessories'] + 6.7*df['Luggage'] + 
                     44.2*df['Swim_Shop'] + 5.7*df['Pioneer_Woman']
                       )/(23.0 + 18.7 + 9.1 + 13.9 + 18.7 + 8.4 + 44.2 + 5.7)

#### 18.2.1.1 Linear Regression Using Statsmodel¶ - without intercept

# Fit and summarize OLS model
mod = sm.OLS(df['Wears-items'], features)

res = mod.fit()

print(res.summary())

#### 18.2.1.2 Linear Regression Using Statsmodel¶ - with intercept

## fit a OLS model with intercept on the data set
X = sm.add_constant(features)

# Fit and summarize OLS model
mod = sm.OLS(df['Wears-items'], X)

res = mod.fit()

print(res.summary())

#### 18.2.2 Linear Regression Using SKLearn

#create numpy array
X = features.values
y = df['Wears-items'].values

#Reshape y
y = y.reshape(-1,1)

#fit regression model
lm = linear_model.LinearRegression()
model = lm.fit(X,y)

#run prediction 
predictions = lm.predict(X)

print((predictions)[0:5])

# generate the score
lm.score(X,y)

lm.coef_

lm.intercept_

### 18.3.1 TARGET - Food-items
Objective: Predict Food Items sales performance

df['Food-items'] = (10.800000*df['Office_supplies '] + 88.600000*df['School_Supplies'] + 44.600000*df['Home_Office'] +
                     11.000000*df['Craft_general'] + 134.3*df['Books'])/(10.800000 + 88.600000 + 44.600000 + 11.000000 + 134.3)

#### 18.3.1.1 Linear Regression Using Statsmodel¶ - without intercept

# Fit and summarize OLS model
mod = sm.OLS(df['Food-items'], features)

res = mod.fit()

print(res.summary())

#### 18.3.1.2 Linear Regression Using Statsmodel¶ - with intercept

# fit a OLS model with intercept on the data set
X = sm.add_constant(features)

# Fit and summarize OLS model
mod = sm.OLS(df['Food-items'], X)

res = mod.fit()

print(res.summary())

#### 18.3.1.2 Linear Regression Using SKLearn

#create numpy array
X = features.values
y = df['Food-items'].values

#Reshape y
y = y.reshape(-1,1)

#fit regression model
lm = linear_model.LinearRegression()
model = lm.fit(X,y)

#run prediction 
predictions = lm.predict(X)

print((predictions)[0:5])

# generate the score
lm.score(X,y)

lm.coef_

lm.intercept_
